+++ lilypad/ee/generations.py
@@ -54,7 +54,7 @@ def specific_generation_version_sync_factory(
 
         return _create_inner_sync(_get_specific_version, sandbox_runner)
 
-    return _specific_generation_version_sync
+    return _specific_generation_version_sync  # pyright: ignore [reportReturnType]
 
 
 def specific_generation_version_async_factory(
@@ -99,4 +99,5 @@ def specific_generation_version_async_factory(
 
         return _create_inner_async(_get_specific_version, sandbox_runner)
 
-    return _specific_generation_version_async
+    return _specific_generation_version_async  # pyright: ignore [reportReturnType]
+
