+++ lilypad/server/_utils/__init__.py
@@ -6,28 +6,10 @@ from .auth import (
     validate_api_key_project_no_strict,
     validate_api_key_project_strict,
 )
-from .spans import (
-    Event,
-    MessageParam,
-    calculate_openrouter_cost,
-    convert_anthropic_messages,
-    convert_events,
-    convert_gemini_messages,
-    convert_mirascope_messages,
-    convert_openai_messages,
-)
 from .versions import construct_function
 
 __all__ = [
     "api_key_header",
-    "Event",
-    "MessageParam",
-    "calculate_openrouter_cost",
-    "convert_anthropic_messages",
-    "convert_events",
-    "convert_gemini_messages",
-    "convert_mirascope_messages",
-    "convert_openai_messages",
     "construct_function",
     "create_api_key",
     "create_jwt_token",
