+++ tests/ee/server/test_require_license.py
@@ -8,8 +8,9 @@ from uuid import UUID, uuid4
 import pytest
 from fastapi import HTTPException, status
 
-from ee import LicenseError, LicenseInfo, Tier
+from ee import LicenseInfo, Tier
 from lilypad.ee.server.require_license import RequireLicense, require_license
+from lilypad.exceptions import LicenseError
 
 
 class DummyProject:
